[convolution] fix MSVC warnings

+++ src/backend/nsl/nsl_conv.c
@@ -68,10 +68,10 @@ int nsl_conv_deconvolution(double s[], size_t n, double r[], size_t m, nsl_conv_
 }
 
 int nsl_conv_linear_direct(double s[], size_t n, double r[], size_t m, nsl_conv_norm_type normalize, nsl_conv_wrap_type wrap, double out[]) {
-	size_t i, j, size = n + m -1, wi = 0;
+	size_t i, j, size = n + m - 1, wi = 0;
 	double norm = 1;
 	if (normalize == nsl_conv_norm_euclidean)
-		norm = cblas_dnrm2(m, r, 1);
+		norm = cblas_dnrm2((int)m, r, 1);
 
 	if (wrap == nsl_conv_wrap_max)
 		nsl_stats_maximum(r, m, &wi);
@@ -79,16 +79,16 @@ int nsl_conv_linear_direct(double s[], size_t n, double r[], size_t m, nsl_conv_
 		wi = m/2;
 
 	for (j = 0; j < size; j++) {
-		int index;
+		int index;	// can be negative
 		double res = 0;
 		for (i = 0; i < n; i++) {
-			index = j - i;
+			index = (int)(j - i);
 			if (index >= 0 && index < (int)m)
 				res += s[i] * r[index]/norm;
 		}
-		index = j - wi;
+		index = (int)(j - wi);
 		if (index < 0)
-			index += size;
+			index += (int)size;
 		out[index] = res;
 	}
 
@@ -99,7 +99,7 @@ int nsl_conv_circular_direct(double s[], size_t n, double r[], size_t m, nsl_con
 	size_t i, j, size = GSL_MAX(n,m), wi = 0;
 	double norm = 1;
 	if (normalize == nsl_conv_norm_euclidean)
-		norm = cblas_dnrm2(m, r, 1);
+		norm = cblas_dnrm2((int)m, r, 1);
 
 	if (wrap == nsl_conv_wrap_max)
 		nsl_stats_maximum(r, m, &wi);
@@ -107,18 +107,18 @@ int nsl_conv_circular_direct(double s[], size_t n, double r[], size_t m, nsl_con
 		wi = m/2;
 
 	for (j = 0; j < size; j++) {
-		int index;
+		int index;	// can be negative
 		double res = 0;
 		for (i = 0; i < n; i++) {
-			index = j-i;
+			index = (int)(j - i);
 			if (index < 0)
-				index += size;
+				index += (int)size;
 			if (index < (int)m)
 				res += s[i]*r[index]/norm;
 		}
-		index = j - wi;
+		index = (int)(j - wi);
 		if (index < 0)
-			index += size;
+			index += (int)size;
 		out[index] = res;
 	}
 
@@ -134,7 +134,7 @@ int nsl_conv_fft_type(double s[], size_t n, double r[], size_t m, nsl_conv_direc
 
 	double norm = 1.;
 	if (normalize == nsl_conv_norm_euclidean)
-		norm = cblas_dnrm2(m, r, 1);
+		norm = cblas_dnrm2((int)m, r, 1);
 
 	if (wrap == nsl_conv_wrap_max)
 		nsl_stats_maximum(r, m, &wi);
